requirements.txt
avro-python3>=1.8.2
pytest>=4.0.1
six>=1.11.0
pylint>=2.4.4
-e ./